infra/lambda_ingest/requirements.txt
opensearch-py==2.7.1
PyYAML==6.0.2
langchain-aws==0.2.6
PyMuPDF==1.25.5
numpy==1.26.4
pillow==10.4.0